requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==18.2.0
Authlib==1.3.2
beautifulsoup4==4.12.3
borb==2.1.25
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
cloudconvert==2.1.0
comtypes==1.4.8
crashtest==0.4.1
cryptography==43.0.3
cssselect2==0.7.0
distlib==0.3.9
distro==1.9.0
docx2pdf==0.1.8
dulwich==0.21.7
fastapi==0.115.5
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.55.2
frozendict==2.4.6
gitignore_parser==0.1.11
google-ads==25.1.0
google-api-core==2.23.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google_search_results==2.4.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
installer==0.7.0
isodate==0.7.2
jaraco.classes==3.4.0
Jinja2==3.1.4
jiter==0.7.1
keyring==24.3.1
lxml==5.3.0
MarkupSafe==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
multitasking==0.0.11
numpy==1.26.4
oauthlib==3.2.2
openai==1.54.4
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdflatex==0.1.3
peewee==3.17.8
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.2
platformdirs==4.3.6
plotly==5.24.1
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
proto-plus==1.25.0
protobuf==5.29.0
ptyprocess==0.7.0
pyairtable==2.3.6
pyarrow==15.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydyf==0.11.0
PyJWT==2.10.0
PyMuPDF==1.24.14
pypandoc==1.14
PyPDF2==3.0.1
pyphen==0.15.0
pyproject_hooks==1.2.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.17
pytrends==4.9.2
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
RapidFuzz==3.10.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
simple-salesforce==1.12.6
six==1.16.0
slack_bolt==1.21.2
slack_sdk==3.33.4
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.2
tqdm==4.66.5
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.1
watchdog==3.0.0
weasyprint==63.0
webencodings==0.5.1
websockets==12.0
writer==0.8.2
writer-sdk==1.5.0
xattr==1.1.0
yfinance==0.2.49
zeep==4.3.1
zipp==3.21.0
zopfli==0.2.3.post1
